# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TodoList API",
    "version": "1.0.0",
    "description": "인증 기반 개인 할일 관리 애플리케이션 API. 사용자는 카테고리와 날짜 범위를 기반으로 할일을 등록·조회·수정·삭제할 수 있습니다."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "로컬 개발 서버"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 액세스 토큰. 로그인(`POST /api/auth/login`) 응답에서 발급된 token 값을 사용합니다."
      }
    },
    "schemas": {
...
      "Todo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "770e8400-e29b-41d4-a716-446655440002"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "example": "660e8400-e29b-41d4-a716-446655440001"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "example": "API 명세 작성"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000,
            "example": "swagger.json 파일 생성 및 OpenAPI 3.0 명세 작성"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2026-05-28"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "종료일은 시작일 이상이어야 합니다 (BR-04: endDate >= startDate)",
            "example": "2026-05-30"
          },
          "status": {
            "type": "string",
            "enum": ["NOT_STARTED", "IN_PROGRESS", "DONE"],
            "description": "허용 상태 전이: NOT_STARTED→IN_PROGRESS, IN_PROGRESS→DONE, IN_PROGRESS→NOT_STARTED, DONE→IN_PROGRESS. 비허용: NOT_STARTED→DONE, DONE→NOT_STARTED (BR-05)",
            "example": "NOT_STARTED"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-28T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-28T00:00:00.000Z"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "message", "code"],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false],
            "example": false
          },
          "message": {
            "type": "string",
            "example": "요청을 처리할 수 없습니다."
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          }
        }
      },
...
      "CreateTodoRequest": {
        "type": "object",
        "required": ["title", "startDate", "endDate"],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "example": "보고서 작성"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000,
            "example": "분기별 성과 보고서 초안 작성"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2026-05-28"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "종료일은 시작일 이상이어야 합니다 (BR-04)",
            "example": "2026-05-31"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "미지정 시 기본 카테고리 자동 적용 (BR-03)",
            "example": "660e8400-e29b-41d4-a716-446655440001"
          }
        }
      },
      "UpdateTodoRequest": {
        "type": "object",
        "description": "수정할 필드만 포함합니다. 하나 이상 포함해야 합니다.",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "example": "보고서 작성 (수정)"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 1000,
            "example": "분기별 성과 보고서 수정본 작성"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2026-05-28"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "종료일은 시작일 이상이어야 합니다 (BR-04)",
            "example": "2026-06-01"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "example": "660e8400-e29b-41d4-a716-446655440001"
          },
          "status": {
            "type": "string",
            "enum": ["NOT_STARTED", "IN_PROGRESS", "DONE"],
            "description": "허용 전이만 가능 (BR-05): NOT_STARTED↔IN_PROGRESS, IN_PROGRESS↔DONE, DONE→IN_PROGRESS. 비허용: NOT_STARTED→DONE, DONE→NOT_STARTED",
            "example": "IN_PROGRESS"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "인증 토큰이 없거나 유효하지 않습니다 (BR-01)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "인증이 필요합니다.",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
      "Forbidden": {
        "description": "접근 권한이 없습니다. 본인 데이터만 접근 가능합니다 (BR-02, BR-07)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "접근 권한이 없습니다.",
              "code": "FORBIDDEN"
            }
          }
        }
      },
      "NotFound": {
        "description": "요청한 리소스를 찾을 수 없습니다",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "리소스를 찾을 수 없습니다.",
              "code": "NOT_FOUND"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/todos": {
      "get": {
        "tags": ["Todos"],
        "summary": "할일 목록 조회",
        "description": "현재 인증된 사용자의 할일 목록을 반환합니다. categoryId·status·overdue·month 쿼리 파라미터로 필터링할 수 있습니다 (BR-01, BR-02). overdue 필터 조건: `endDate < 오늘 AND status != DONE`.",
        "operationId": "getTodos",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "description": "카테고리 UUID로 필터링. 해당 카테고리에 속한 할일만 반환합니다.",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "660e8400-e29b-41d4-a716-446655440001"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "상태로 필터링 (NOT_STARTED | IN_PROGRESS | DONE)",
            "schema": {
              "type": "string",
              "enum": ["NOT_STARTED", "IN_PROGRESS", "DONE"]
            },
            "example": "IN_PROGRESS"
          },
          {
            "name": "overdue",
            "in": "query",
            "required": false,
            "description": "true로 설정 시 기한 초과 미완료 할일만 반환합니다 (`endDate < 오늘 AND status != DONE`). status 필터와 함께 사용 시 overdue 조건이 우선 적용됩니다.",
            "schema": {
              "type": "boolean"
            },
            "example": true
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "description": "YYYY-MM 형식으로 지정 시 해당 월의 start_date를 가진 할일만 반환합니다. 다른 필터와 AND 조건으로 결합됩니다.",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}$"
            },
            "example": "2026-05"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 목록 조회 성공. 필터 조건에 맞는 할일 배열을 반환합니다. 해당하는 할일이 없으면 빈 배열을 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "todos": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Todo"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "todos": [
                      {
                        "id": "770e8400-e29b-41d4-a716-446655440002",
                        "userId": "550e8400-e29b-41d4-a716-446655440000",
                        "categoryId": "660e8400-e29b-41d4-a716-446655440001",
                        "title": "보고서 작성",
                        "description": "분기별 성과 보고서",
                        "startDate": "2026-05-28",
                        "endDate": "2026-05-31",
                        "status": "IN_PROGRESS",
                        "createdAt": "2026-05-28T00:00:00.000Z",
                        "updatedAt": "2026-05-28T00:00:00.000Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Todos"],
        "summary": "할일 등록",
        "description": "새 할일을 생성합니다 (BR-01, BR-03, BR-04). 제목·시작일·종료일은 필수입니다. categoryId 미지정 시 기본 카테고리가 자동 적용됩니다 (BR-03). 종료일은 시작일 이상이어야 합니다 (BR-04).",
        "operationId": "createTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoRequest"
              },
              "examples": {
                "withCategory": {
                  "summary": "카테고리 지정",
                  "value": {
                    "title": "보고서 작성",
                    "description": "분기별 성과 보고서 초안 작성",
                    "startDate": "2026-05-28",
                    "endDate": "2026-05-31",
                    "categoryId": "660e8400-e29b-41d4-a716-446655440002"
                  }
                },
                "withoutCategory": {
                  "summary": "카테고리 미지정 (기본 카테고리 자동 적용)",
                  "value": {
                    "title": "장보기",
                    "startDate": "2026-05-28",
                    "endDate": "2026-05-28"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "할일 등록 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "todo": {
                          "$ref": "#/components/schemas/Todo"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "todo": {
                      "id": "770e8400-e29b-41d4-a716-446655440003",
                      "userId": "550e8400-e29b-41d4-a716-446655440000",
                      "categoryId": "660e8400-e29b-41d4-a716-446655440001",
                      "title": "장보기",
                      "description": null,
                      "startDate": "2026-05-28",
                      "endDate": "2026-05-28",
                      "status": "NOT_STARTED",
                      "createdAt": "2026-05-28T03:00:00.000Z",
                      "updatedAt": "2026-05-28T03:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 유효성 오류 — 제목 미입력 또는 날짜 유효성 위반 (BR-04)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingTitle": {
                    "summary": "제목 미입력",
                    "value": {
                      "success": false,
                      "message": "할일 제목은 필수입니다.",
                      "code": "MISSING_TODO_TITLE"
                    }
                  },
                  "invalidDate": {
                    "summary": "날짜 유효성 위반 (BR-04)",
                    "value": {
                      "success": false,
                      "message": "종료일은 시작일 이후여야 합니다.",
                      "code": "INVALID_DATE_RANGE"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/todos/{id}": {
      "get": {
        "tags": ["Todos"],
        "summary": "할일 단건 조회",
        "description": "지정한 할일의 상세 정보를 반환합니다. 본인 소유 할일만 조회할 수 있습니다 (BR-01, BR-02).",
        "operationId": "getTodoById",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "조회할 할일 UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "770e8400-e29b-41d4-a716-446655440002"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 단건 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "todo": {
                          "$ref": "#/components/schemas/Todo"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": ["Todos"],
        "summary": "할일 수정",
        "description": "지정한 할일을 수정합니다 (BR-02, BR-04, BR-05). 수정할 필드만 포함합니다. 종료일은 시작일 이상이어야 합니다 (BR-04). 상태 변경 시 허용된 전이만 가능합니다 (BR-05): NOT_STARTED↔IN_PROGRESS, IN_PROGRESS↔DONE, DONE→IN_PROGRESS. 비허용 전이: NOT_STARTED→DONE, DONE→NOT_STARTED.",
        "operationId": "updateTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "수정할 할일 UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "770e8400-e29b-41d4-a716-446655440002"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoRequest"
              },
              "examples": {
                "updateStatus": {
                  "summary": "상태 변경 (BR-05 허용 전이)",
                  "value": {
                    "status": "IN_PROGRESS"
                  }
                },
                "updateTitleAndDate": {
                  "summary": "제목 및 날짜 변경",
                  "value": {
                    "title": "보고서 작성 (수정)",
                    "endDate": "2026-06-01"
                  }
                },
                "updateCategory": {
                  "summary": "카테고리 변경",
                  "value": {
                    "categoryId": "660e8400-e29b-41d4-a716-446655440002"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "할일 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "todo": {
                          "$ref": "#/components/schemas/Todo"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 유효성 오류 — 날짜 유효성 위반 (BR-04) 또는 허용되지 않는 상태 전이 (BR-05)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidDate": {
                    "summary": "날짜 유효성 위반 (BR-04)",
                    "value": {
                      "success": false,
                      "message": "종료일은 시작일 이후여야 합니다.",
                      "code": "INVALID_DATE_RANGE"
                    }
                  },
                  "invalidStatusTransition": {
                    "summary": "허용되지 않는 상태 전이 (BR-05)",
                    "value": {
                      "success": false,
                      "message": "허용되지 않는 상태 전이입니다. (NOT_STARTED에서 DONE으로 직접 변경 불가)",
                      "code": "INVALID_STATUS_TRANSITION"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "할일 삭제",
        "description": "지정한 할일을 영구 삭제합니다. 본인 소유 할일만 삭제할 수 있습니다 (BR-01, BR-02).",
        "operationId": "deleteTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "삭제할 할일 UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "770e8400-e29b-41d4-a716-446655440002"
          }
        ],
        "responses": {
          "200": {
            "description": "할일 삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  }
}